pyproject.toml
[project]
name = "glados-checkin"
version = "1.0.0"
description = "GLaDOS auto checkin script"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "playwright>=1.48.0",
]

[tool.uv]
dev-dependencies = []
